--- a/AUTHORS
+++ b/AUTHORS
@@ -17,3 +17,4 @@ Contributors
 - Byron Roosa
 - Andrew Crozier
 - @eight04
+- Martin Di Paola @eldipa