--- a/benchmark.py
+++ b/benchmark.py
@@ -10,6 +10,10 @@
         .....................
         ls.input: Mean +- std dev: 644 ns +- 23 ns
 
+        $ BENCHMARK=tests/captured/ls.input GEOMETRY=1024x1024 python benchmark.py -o results.json
+        .....................
+        ls.input: Mean +- std dev: 644 ns +- 23 ns
+
     Environment variables:
 
     BENCHMARK: the input file to feed pyte's Stream and render on the Screen
@@ -32,26 +36,54 @@ except ImportError:
 
 import pyte
 
-
-def make_benchmark(path, screen_cls, columns, lines):
+def setup(path, screen_cls, columns, lines):
     with io.open(path, "rt", encoding="utf-8") as handle:
         data = handle.read()
 
-    stream = pyte.Stream(screen_cls(columns, lines))
+    screen = screen_cls(columns, lines)
+    stream = pyte.Stream(screen)
+
+    return data, screen, stream
+
+def make_stream_feed_benchmark(path, screen_cls, columns, lines):
+    data, _, stream = setup(path, screen_cls, columns, lines)
     return partial(stream.feed, data)
 
+def make_screen_display_benchmark(path, screen_cls, columns, lines):
+    data, screen, stream = setup(path, screen_cls, columns, lines)
+    stream.feed(data)
+    return lambda: screen.display
+
+def make_screen_reset_benchmark(path, screen_cls, columns, lines):
+    data, screen, stream = setup(path, screen_cls, columns, lines)
+    stream.feed(data)
+    return screen.reset
+
+def make_screen_resize_half_benchmark(path, screen_cls, columns, lines):
+    data, screen, stream = setup(path, screen_cls, columns, lines)
+    stream.feed(data)
+    return partial(screen.resize, lines=lines//2, columns=columns//2)
 
 if __name__ == "__main__":
     benchmark = os.environ["BENCHMARK"]
     lines, columns = map(int, os.environ.get("GEOMETRY", "24x80").split('x'))
     sys.argv.extend(["--inherit-environ", "BENCHMARK,GEOMETRY"])
 
-    runner = Runner(metadata={
+    runner = Runner()
+
+    metadata = {
         'input_file': benchmark,
         'columns': columns,
         'lines': lines
-        })
+        }
 
+    benchmark_name = os.path.basename(benchmark)
     for screen_cls in [pyte.Screen, pyte.DiffScreen, pyte.HistoryScreen]:
-        name = os.path.basename(benchmark) + "->" + screen_cls.__name__
-        runner.bench_func(name, make_benchmark(benchmark, screen_cls, columns, lines))
+        screen_cls_name = screen_cls.__name__
+        for make_test in (make_stream_feed_benchmark, make_screen_display_benchmark, make_screen_reset_benchmark, make_screen_resize_half_benchmark):
+            scenario = make_test.__name__[5:-10] # remove make_ and _benchmark
+
+            name = f"[{scenario} {lines}x{columns}] {benchmark_name}->{screen_cls_name}"
+            metadata.update({'scenario': scenario, 'screen_cls': screen_cls_name})
+            runner.bench_func(name, make_test(benchmark, screen_cls, columns, lines), metadata=metadata)
+