… tests/helpers/asserts.py
… tests/test_diff.py
… tests/test_history.py
… tests/test_input_output.py
… tests/test_screen.py
… tests/test_stream.py