@eldipa 2022-07-16
@superbobry , the PR is quite long but I have no problem to review it with you, piece by piece. Despite its length I think that the changes are an solid improvement (ok, I'm a little biased here).

A possible first round for the review process could be:

 - review [8513fe8](https://github.com/selectel/pyte/pull/160/commits/8513fe82c586530e5fc361154ef974702e3ef116) to [eec4a2e](https://github.com/selectel/pyte/pull/160/commits/eec4a2ee0afbb31132c0cba689d49bf068533ddc): changes on how the benchmark was done
 - from there to [020fce6](https://github.com/selectel/pyte/pull/160/commits/020fce61c2f5eb5467727048743e481b2a8d2a87) for `Screen.display` optimizations
 - from there to  [9721698](https://github.com/selectel/pyte/pull/160/commits/9721698a3132f67ee78c976054d847237be973ca) for `Screen.draw` optimizations

This walkthrough covers ~30% of the PR. From there we can think in the next rounds for review.