@Moult 2024-12-23
Bump - any chance of reviewing this? I've also been experiencing performance issues (I'm using feed). It's 26 commits behind but it looks pretty easy to sync up.