@eldipa 2024-12-23
> Bump - any chance of reviewing this? I've also been experiencing performance issues (I'm using feed). It's 26 commits behind but it looks pretty easy to sync up.

Hi @Moult , I'm ok to assist in the review of the PR but I'm not sure if there is intention to merge it by the owners of the project. Until that, doing a rebase + fixing the conflict may yield no result.

If you want, you could try to use the PR directly to see if there is an improvement in the performance or not. Depending of your use case, you may also try a fork that I made where I removed features in order to be a little faster.

PR branch (compatible with the functionality of `pyte`): https://github.com/byexamples/pyte/tree/Screen-Optimizations
Fork (with less features but faster): https://github.com/byexamples/termscraper