@Moult 2024-12-23
Thanks! I've actually tested it and I can confirm there are really, really significant speedups to feed(), and as far as I can tell (at least in the realm of NetHack ttyrecs) there are no regressions. I don't use display so I can't really commment there.

One minor (major?) issue for me is that the Char namedtuple has changed into a class. I write a ttyrec player, and so to implement a "rewind" feature, I copy the contents of buffer (or now, since there is a BufferView, I copy _buffer). The structure of buffer used to be basically dict[dict[namedtuple]] whereas now it's dict[dict[class[namedtuple, namedtuple]]] and that absolutely kills performance if one needed to do a copy of buffer 500,000 times like myself.

I naively reintroduced the Char namedtuple and tweaked write_data to use it and it seems to work fine again, and now super fast buffer copying via this code works. I'd love to know if this has side effects or why a Char class was chosen:

```
# Example copy which is only fast if Char is a namedtuple, and not possible if Char is a class
buffer_copy = {y: dict(row) for y, row in self.screen._buffer.items()}
```